feat: added stories for TextArea component --storybook
